added the typescript dependency rather than fighting with es5
