+++ .npmignore
@@ -1,0 +1,5 @@
+src
+tsconfig.json
+tslint.json
+.prettierrc
+.jestconfig.json
