+++ index.js
@@ -1,7 +1,16 @@
 const chalk = require('chalk');
 const figlet = require('figlet');
 const package = require('./package.json');
- 
+
+//libs
+// const AppolloController = require('./lib/ApolloController');
+const Model = require('./lib/Model');
+const Animal = require('./lib/Animal');
+// import Model from './lib/Model';
+// const Molecule = require('./lib/Molecule');
+// const Muon = require('./lib/Muon');
+
+
 module.exports.init = () => {
   console.log(
     chalk.green(
@@ -19,4 +28,5 @@ module.exports.version = () => {
   return '';
 }
 
-module.exports.
+
+//
