+++ lib/Animal.js
@@ -1,0 +1,7 @@
+class Animal {
+  constructor(){
+    this.name = "Cow";
+  }
+}
+
+module.exports = Animal;
