+++ lib/index.d.ts
@@ -1,0 +1,1 @@
+export declare const Hello: (name: String) => void;
