+++ output/index.js
@@ -1,0 +1,7 @@
+'use strict';
+Object.defineProperty(exports, '__esModule', { value: true });
+exports.Hello = void 0;
+const Hello = (name) => {
+  console.log(`Hello ${name}`);
+};
+exports.Hello = Hello;
