+++ src/__tests__/Hello.test.ts
@@ -1,0 +1,5 @@
+import {Hello} from "../index";
+
+test('Hello', () => {
+  expect(Hello('DiamondHands')).toBe('Hello DiamondHands');
+});
