+++ src/index.ts
@@ -1,0 +1,3 @@
+export const Hello = (name: String) => {
+  console.log(`Hello ${name}`);
+}
